+++ API/Models/BankDTOs/BankAccountDTO.cs
@@ -14,8 +14,8 @@ public class BankAccountDTO
     public AppUser AppUser { get; set; }
 
     public string HolderName { get; set; }
-    public double SavingsBalance { get; set; }
-    public double CheckingsBalance { get; set; }
+    public decimal SavingsBalance { get; set; }
+    public decimal CheckingsBalance { get; set; }
     [Range(1, 500, ErrorMessage = "Amount must be between 1 and 500.")]
-    public double TransactionAmount { get; set; }
+    public decimal TransactionAmount { get; set; }
 }
