+++ Client/Models/OrderFood/OrderHeader.cs
@@ -14,7 +14,7 @@ public class OrderHeader
     [ValidateNever, ForeignKey(nameof(AppUserId))]
     public AppUser AppUser { get; set; }
 
-    public double OrderTotal { get; set; }
+    public decimal OrderTotal { get; set; }
     public string TrackingNumber { get; set; }
     public DateTime OrderDate { get; set; }
